Add the Trame UI skeleton
