+++ src/NeuXtalViz/qt/views/experiment_planner.py
@@ -32,7 +32,7 @@ from matplotlib.backends.backend_qtagg import NavigationToolbar2QT
 from matplotlib.figure import Figure
 from matplotlib.ticker import FormatStrFormatter
 
-from NeuXtalViz.views.base_view import NeuXtalVizWidget
+from NeuXtalViz.qt.views.base_view import NeuXtalVizWidget
 
 
 class ExperimentView(NeuXtalVizWidget):
