+++ src/NeuXtalViz/qt/views/periodic_table.py
