+++ src/NeuXtalViz/qt/views/reduction_plan.py
