+++ src/NeuXtalViz/qt/views/utilities.py
